Image data bucket only allows integer data.

+++ pyxel/models/readout_electronics/sar_adc.py
@@ -55,7 +55,7 @@ def apply_sar_adc(
         # Divide reference voltage by 2 for next step
         ref /= 2.0
 
-    return data_digitized_2d
+    return data_digitized_2d.astype(np.uint64)
 
 
 # TODO: documentation, range volt - only max is used
